backend/requirements.txt
# Flask核心框架
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
flask-restx==1.1.0

# Web服务器
werkzeug==2.3.7

# 数据库相关
psycopg2-binary==2.9.7
SQLAlchemy==2.0.19

# 调度任务
APScheduler==3.10.1

# HTTP客户端
requests==2.31.0

# JWT认证
Flask-JWT-Extended==4.5.2

# 环境配置
python-dotenv==1.0.0

# 日志
colorlog==6.7.0

# 序列化
marshmallow==3.20.1

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 中文分词
jieba==0.42.1

# 数据处理和导出 (修复版本兼容性)
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2

# 开发工具
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.1.0

# 生产服务器
gunicorn==21.2.0 